Moved some code around and fixed pipeline bug
